Mark all methods as pure
This means every function can be called from an immutable / pure context

Note: ToArray will fail psalm on this commit as (rightfully) there is an impure psalm issue

This commit is a bit of a lie to psalm: Psalm is correct that our object is not actually immutable / pure.

But is if you remove caching.

This lie seems tolerable (despite how bad a practice overriding the typechecker is) because the mutation is entirely memoization.