ignore private constants
Update use of reflection to only populate the cache with public and
protected constants, ie. those that would be visible to the abstract
class if not for the use of reflection. Private constants will now be
treated as invalid - they no longer appear in toArray(), can't be
instantiated using static method calls, and throw an exception as normal
when constructing the enum with the private value hard-coded.

Resolves #120.