Added new type-safe API that allows validating a given value against a given `class-string<Enum>`
This patch introduces two new methods:

 * `Enum::isValidEnumValue()`
 * `Enum::assertValidEnumValue()`

The `Enum::isValidEnumValue()` is still wonky due to https://github.com/vimeo/psalm/issues/5372, but
overall, this allows for matching against other `Enum` sub-types, by validating that a given `$value`
is effectively a subtype of `T` within an `Enum<T>`.

The previous approach did **not** work, because static method types are not coupled with the class they
are put on (they are effectively namespaced functions, and not much else).

In here, we also:
 * completed test coverage for `Enum::from()`
 * pinned `vimeo/psalm:4.6.2`, mostly for build stability
 * added coverage mapping for `phpunit.xml`